+++ s4e-backend/src/main/java/pl/cyfronet/s4e/bean/EmailVerification.java
@@ -8,6 +8,9 @@ import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
+/**
+ * There is ON CASCADE DELETE on Appuser
+ */
 @Entity
 @Data
 @Builder