+++ s4e-backend/src/main/resources/db/migration/V25__add_fk_constraint_app_user_to_email_verification_table.sql
@@ -1,0 +1,6 @@
+ALTER TABLE email_verification
+DROP CONSTRAINT email_verification_app_user_id_fkey,
+ADD CONSTRAINT email_verification_app_user_id_fkey
+FOREIGN KEY (app_user_id)
+REFERENCES app_user(id)
+ON DELETE CASCADE;