+++ CHANGELOG.md
@@ -11,5 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Possibility to set maven artifact version with properties
 - This changelog file
+- migration with 'On delete cascade' for AppUser field on PasswordReset and EmailVerification tables
+
+### Fixed
+
+- Constraints names on Product and Scene tables
 
 [unreleased]: https://github.com/cyfronet-fid/sat4envi/compare/0ebb1138...HEAD